@@ -2,20 +2,16 @@ apply plugin: 'com.android.library'
 apply plugin: 'kotlin-android'
 apply plugin: 'kotlin-android-extensions'
 
-apply plugin: 'com.github.dcendents.android-maven'
-
 group = 'com.github.DonMat'
 
 android {
-    compileSdkVersion 27
+    compileSdkVersion 33
 
 
 
     defaultConfig {
         minSdkVersion 21
-        targetSdkVersion 27
-        versionCode 1
-        versionName "1.0.1"
+        targetSdkVersion 33
     }
 
     buildTypes {
@@ -24,13 +20,13 @@ android {
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
         }
     }
+    namespace 'pl.utkala.searchablespinner'
 
 }
 
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
-    implementation 'com.android.support:appcompat-v7:27.1.1'
-    implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
+    implementation 'androidx.appcompat:appcompat:1.5.1'
 }
 repositories {
     mavenCentral()
